some-of-the-backend-scripts/requirements (1).txt
bleach==3.3.0
Flask==3.0.2
mysql-connector-python==8.3.0
